[go-gen] Refactor go code generation by package
